disentangle engine to a file
